% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_lab_gen.R
\name{create_pdf}
\alias{create_pdf}
\title{Generate a Complete LaTeX Document and Compile it into a PDF}
\usage{
create_pdf(
  file_out,
  ind_list,
  print_info,
  lab_size = 15,
  n_col = 8,
  col_N_name = "N",
  hl_col = "orange"
)
}
\arguments{
\item{file_out}{A character string specifying the path to the LaTeX file that will be generated
and compiled into a PDF.}

\item{ind_list}{A data frame containing individual data. Each row represents data
for one label to be printed in the LaTeX document.}

\item{print_info}{A data frame specifying the parameters for printing, including field names,
formatting options, and what data to print.}

\item{lab_size}{An integer specifying the size (in mm) for the label (default is 15 mm).}

\item{n_col}{An integer specifying the number of label columns in the LaTeX document (default is 8 columns).}

\item{col_N_name}{A character string specifying the column name to be used for labels duplication (default is "N").}

\item{hl_col}{A character string specifying the highlight color for specific elements (default is "orange").}
}
\value{
The function creates a LaTeX file, compiles it into a PDF, and saves the output
to the specified location. It does not return any value in R.
}
\description{
This function generates a LaTeX document by sequentially adding a header, printing
labels for each line of data, appending a footer, and then compiling the LaTeX code
into a PDF document.
}
\details{
This function first calls \code{print_header} to write the beginning of the LaTeX
document. Then, it iterates over each row of \code{ind_list}, calling \code{print_line}
to generate labels according to the provided \code{print_info}. After all labels are printed,
it appends the LaTeX document footer using \code{print_bottom}. Finally, it compiles the
LaTeX document into a PDF using \code{pdflatex}.
}
\examples{
# Example usage:
# create_pdf("output.tex", ind_list, print_info, lab_size = 20, n_col = 10)

}
