% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_lab_gen.R
\name{print_header}
\alias{print_header}
\title{Create LaTeX Header for mosquito labels Document}
\usage{
print_header(file_out, lab_size = 15, n_col = 8)
}
\arguments{
\item{file_out}{A character string specifying the name of the output file to write the LaTeX header to.}
}
\description{
This function generates a LaTeX document header for mosquito labels Document
and writes it to the specified output file. The header includes various LaTeX packages,
document settings, metadata and the document front cover. The latex code is based on the work
by Samuel Brown (see https://github.com/sdjbrown/publicFiles/blob/master/labels.tex and
http://the-praise-of-insects.blogspot.com/2010/03/latex-insect-labels.html )
}
\examples{
\dontrun{
print_header("output.tex")
}

}
