% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_lab_gen.R
\name{InsectLabelR}
\alias{InsectLabelR}
\title{Launch the InsectLabelR Shiny application}
\usage{
InsectLabelR()
}
\description{
This function launches an interactive Shiny application allowing users
to use EtiquetteR with either example datasets included in the package
or user-provided data.
}
